+++ stress/v2/statement/query.go
@@ -142,7 +142,11 @@ func getRandomFieldKey(m map[string]interface{}) string {
 }
 
 func setMapValues(m map[string]interface{}, p models.Point) {
-	m["%f"] = getRandomFieldKey(p.Fields())
+	fields, err := p.Fields()
+	if err != nil {
+		panic(err)
+	}
+	m["%f"] = getRandomFieldKey(fields)
 	m["%m"] = p.Name()
 	m["%t"] = getRandomTagPair(p.Tags())
 	m["%a"] = p.UnixNano()